requirements.txt
telethon==1.34.0
numpy>=1.24.0
pandas==2.0.2
vaderSentiment==3.3.2
streamlit==1.28.0
python-dotenv==1.0.1
matplotlib==3.8.0
wordcloud==1.9.3
seaborn==0.13.0
openai==0.28.1
python-telegram-bot==20.5
aiogram>=3.0
